feat: Implement initial Next.js project structure with arcade UI components and game physics utilities.
