+++ app/package.json
@@ -24,6 +24,7 @@
     "format": "prettier --write \"**/*.{js,jsx,json,md}\"",
     "format:check": "prettier --check \"**/*.{js,jsx,json,md}\"",
     "lighthouse": "node ./scripts/run-lighthouse.js",
+    "type-check": "tsc --noEmit",
     "prepare": "cd .. && husky app/.husky"
   },
   "devDependencies": {
@@ -34,6 +35,9 @@
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^14.6.1",
+    "@types/node": "^24.10.1",
+    "@types/react": "^19.2.6",
+    "@types/react-dom": "^19.2.3",
     "@typescript-eslint/parser": "^8.46.2",
     "babel-plugin-styled-components": "^2.1.4",
     "chrome-launcher": "^1.2.1",
@@ -49,6 +53,7 @@
     "jest-environment-jsdom": "^30.2.0",
     "lighthouse": "^13.0.1",
     "lint-staged": "^16.2.6",
-    "prettier": "^3.6.2"
+    "prettier": "^3.6.2",
+    "typescript": "^5.9.3"
   }
 }
