+++ docs/TODO-old.md
@@ -1,0 +1,59 @@
+# TODO
+
+## Phase 10 - ASTEROID GAME ✅ COMPLETE
+
+See `PHASE_10_PLAN.md` for full implementation details.
+
+### Asteroid Game - Sprint 1-8 ✅ COMPLETE
+
+- [x] **Weapon Visual Effects** - Muzzle flashes, shell casings, impact effects ✅
+- [x] **Target Visual Feedback** - Spawn/hit/split/destroy animations with particles ✅
+- [x] **Performance Optimization** - Object pooling (planned for future), spatial partitioning ✅
+- [x] **Audio Polish** - Spatial sound, layered impacts, dynamic music system ✅
+- [x] **Settings Menu** - Mouse sensitivity, accessibility options (colorblind, reduce motion, high contrast) ✅
+- [x] **Dynamic Music** - 4-layer procedural audio with wave progression ✅
+- [x] **Kill Streak System** - Visual announcements + power chord sounds ✅
+- [x] **Screen Effects** - Low health vignette, screen shake, proximity warnings ✅
+- [x] **E2E Testing** - Reliable CI/CD pipeline with Playwright ✅
+
+### Future Enhancements (Phase 11+)
+
+- [ ] **CRT Effect Polish** - Thinner scanlines, subtle jitter
+- [ ] **Wave Mechanics** - Runner targets, AI targets that shoot back, boss waves, randomized power-up drops
+- [ ] **Challenge Modes** - Time attack, no-damage runs, boss rush
+- [ ] **Object Pooling** - InstancedMesh for hit particles (partially done, needs expansion)
+
+### Code Quality
+
+- [ ] Test Coverage expansion (unit tests for shared systems)
+- [ ] JSDoc documentation for shared systems
+- [ ] Performance profiling and optimization
+
+---
+
+## Phase 11 - NEW GAMES 📋 PLANNING
+
+See "New Games Roadmap" section in `PHASE_10_PLAN.md`.
+
+### Framework Reuse Strategy
+
+- **Shared Audio System** ✅ Ready (AudioManager, SoundManager, DynamicMusicSystem)
+- **Shared Input System** ✅ Ready (KeyboardManager, MouseManager)
+- **Shared Scoring System** ✅ Ready (ScoreManager, HighScoreManager, StatsTracker)
+- **Shared UI Components** 🔄 Needs extraction (ArcadeButton, ArcadeMenu, ArcadeCard from existing components)
+- **Shared Physics** 🔄 Needs extraction (collision detection, spatial grid from Asteroid)
+
+### Quick Win Games (1-2 Days Each)
+
+1. **Breakout/Arkanoid** 🧱 - 2D paddle physics, brick patterns (LOW complexity)
+2. **Space Invaders** 👾 - Formation movement, enemy patterns (MEDIUM complexity)
+3. **Flappy Bird Clone** 🐦 - Simple 2D physics, endless runner (LOW complexity)
+4. **Pong 3D** 🏓 - AI opponent, 3D perspective (LOW complexity)
+
+### Target: 3-5 Playable Games
+
+- ✅ **Asteroid** (6DOF space shooter) - COMPLETE with full polish
+- 🔄 **FPS** (terrain-based 3D) - In progress
+- 📋 **Breakout** - Next quick win (recommended)
+- 📋 **Space Invaders** - Pattern AI showcase
+- 📋 **Flappy Bird** - Mobile-ready endless runner
